+++ src/rdm/parameters.c
@@ -1,7 +1,8 @@
 #include "parameters.h"
 
+#include "dmx/hal.h"
+#include "dmx/struct.h"
 #include "esp_dmx.h"
-#include "dmx/driver.h"
 #include "rdm/utils.h"
 
 static const char *TAG = "rdm_parameter";  // The log tagline for the file.
@@ -18,19 +19,19 @@ bool rdm_get_device_info(dmx_port_t dmx_num, rdm_device_info_t *device_info) {
 bool rdm_get_software_version_label(dmx_port_t dmx_num,
                                     char *software_version_label,
                                     size_t *size) {
-    DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
-    DMX_CHECK(software_version_label != NULL, 0,
-              "software_version_label is null");
-    DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(software_version_label != NULL, 0,
+            "software_version_label is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
 
-    return rdm_get_parameter(dmx_num, RDM_PID_SOFTWARE_VERSION_LABEL,
-                             software_version_label, size);
+  return rdm_get_parameter(dmx_num, RDM_PID_SOFTWARE_VERSION_LABEL,
+                           software_version_label, size);
 }
 
 bool rdm_get_identify_device(dmx_port_t dmx_num, uint8_t *identify) {
-    DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
-    DMX_CHECK(identify != NULL, 0, "identify is null");
-    DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
+  DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
+  DMX_CHECK(identify != NULL, 0, "identify is null");
+  DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
 
   size_t size = sizeof(*identify);
   return rdm_get_parameter(dmx_num, RDM_PID_IDENTIFY_DEVICE, identify, &size);
@@ -50,7 +51,7 @@ bool rdm_get_dmx_start_address(dmx_port_t dmx_num,
   DMX_CHECK(dmx_num < DMX_NUM_MAX, 0, "dmx_num error");
   DMX_CHECK(dmx_start_address != NULL, 0, "dmx_start_address is null");
   DMX_CHECK(dmx_driver_is_installed(dmx_num), 0, "driver is not installed");
-  
+
   // DMX getter can be called and should always return true
   *dmx_start_address = dmx_get_start_address(dmx_num);
   return (dmx_start_address != 0);
