+++ src/rdm/parameters.h
@@ -15,11 +15,11 @@ extern "C" {
 
 /**
  * @brief Gets a copy of the RDM device info of this device.
- * 
+ *
  * @param dmx_num The DMX port number.
- * @param[out] device_info A pointer which stores a copy of the device info of 
+ * @param[out] device_info A pointer which stores a copy of the device info of
  * this device.
- * @return true on success. 
+ * @return true on success.
  * @return false on failure.
  */
 bool rdm_get_device_info(dmx_port_t dmx_num, rdm_device_info_t *device_info);
@@ -40,22 +40,22 @@ bool rdm_get_software_version_label(dmx_port_t dmx_num,
 
 /**
  * @brief Gets a copy of the RDM identify device state of this device.
- * 
+ *
  * @param dmx_num The DMX port number.
  * @param identify A pointer which stores a copy of the identify device state of
  * this device.
- * @return true on success. 
+ * @return true on success.
  * @return false on failure.
  */
 bool rdm_get_identify_device(dmx_port_t dmx_num, uint8_t *identify);
 
 /**
  * @brief Sets the RDM identify device state of this device.
- * 
+ *
  * @param dmx_num The DMX port number.
  * @param identify The identify device state to which to set this device.
  * @return true on success.
- * @return false on failure. 
+ * @return false on failure.
  */
 bool rdm_set_identify_device(dmx_port_t dmx_num, const uint8_t identify);
