+++ src/rdm/utils.c
@@ -4,7 +4,8 @@
 #include <string.h>
 
 #include "dmx/caps.h"
-#include "dmx/driver.h"
+#include "dmx/hal.h"
+#include "dmx/struct.h"
 #include "endian.h"
 #include "esp_dmx.h"
 #include "esp_log.h"
